rename: 패키지 이름 변경, 이동
리포지토리들 repository 패키지로 이동
auditing -> base
